Extract a Header component
